Update README.md
A resource block in a terraform code will create only one real infrastructure. If we want to create more resource (E.g., EC2) we have to create multiple resource block. But this is not an efficient way. For that we can use count argument. 
The count argument will accept a whole number and create that many instances of the resource we want. 
While using count argument we can also make use of count.index which holds a distinct index number, starting from 0 that uniquely identifies each instance created by the count meta argument